TimeSample attributes changed to float, Madgwick and calculation of gloabal accelerations added

+++ app/src/main/java/com/example/gsurfexample/TimeSample.java
@@ -9,18 +9,18 @@ public class TimeSample {
     @PrimaryKey(autoGenerate = true)
     private int id;
     private long timeStamp;
-    private double ddx, ddy, ddz;
-    private double gFx, gFy, gFz;
-    private double Bx, By, Bz;
-    private double wx, wy, wz;
+    private float ddx, ddy, ddz;
+    private float gFx, gFy, gFz;
+    private float Bx, By, Bz;
+    private float wx, wy, wz;
     private double lat, lon, height;
     private double xGPS, yGPS;
 
 
-    public TimeSample(long timeStamp, double ddx, double ddy, double ddz,   // must be like that due to ROOM
-                      double gFx, double gFy, double gFz,
-                      double Bx, double By, double Bz,
-                      double wx, double wy, double wz,
+    public TimeSample(long timeStamp, float ddx, float ddy, float ddz,   // must be like that due to ROOM
+                      float gFx, float gFy, float gFz,
+                      float Bx, float By, float Bz,
+                      float wx, float wy, float wz,
                       double lat, double lon, double height,
                       double xGPS, double yGPS) {
 
@@ -55,50 +55,50 @@ public class TimeSample {
 
     public long getTimeStamp() { return timeStamp; }
 
-    public double getDdx() {
+    public float getDdx() {
         return ddx; }
 
-    public double getDdy() {
+    public float getDdy() {
         return ddy;
     }
 
-    public double getDdz() {
+    public float getDdz() {
         return ddz;
     }
 
-    public double getGFx() {
+    public float getGFx() {
         return gFx;
     }
 
-    public double getGFy() {
+    public float getGFy() {
         return gFy;
     }
 
-    public double getGFz() {
+    public float getGFz() {
         return gFz;
     }
 
-    public double getBx() {
+    public float getBx() {
         return Bx;
     }
 
-    public double getBy() {
+    public float getBy() {
         return By;
     }
 
-    public double getBz() {
+    public float getBz() {
         return Bz;
     }
 
-    public double getWx() {
+    public float getWx() {
         return wx;
     }
 
-    public double getWy() {
+    public float getWy() {
         return wy;
     }
 
-    public double getWz() {
+    public float getWz() {
         return wz;
     }
